Active sched entity tracking
In preperation for period agnostic tracing, track the active sched
enitities when there an enqueue. Although We don't remove these entities
from our active queue on a dequeue. This is because entities may get
enqueued and dequeued constantly due to slice expiry and we want to
ensure that all relevant entities are accounted for when we actually
attempt to make a recommendation. The active sched enity list is cleared
when a recommendation is made.

This may lead to a case where some sched entities that were dequeued
never get enqueued again within that period of time, leading slightly
higher recommendations. However, to cater for the most common case of
being rescheduled, incuring a minor CPU overhead temporarily seems like
a good trade-off.

Signed-off-by: Pratik Rajesh Sampat <psampat2@illinois.edu>