sched: Period-agnostic runtime tracing
Trace runtime and yield information across the boundary of a bandwidth
period by priming two clocks runtime and yieldtime clock. The runtime
clock keeps ticking until a legitimate yield is found. The yieldtime
clock gets reprimed at every yeild wheather it was caused by the
application itself or by other factors such as sched_slice expiry.

A self-yeild is identified if the duration of the yeild was more than
than of the last sched_slice. Of course, this can also be triggered by
other factors such as throttle - which will be addressed in future
patches.

Signed-off-by: Pratik Rajesh Sampat <psampat2@illinois.edu>