sched: Support for exponential scaling
During high degrees of throttle, the workload behavior may change and
can cause the period agnostic and bound algorithms to incorrectly
identify behavior. Therefore, when the throttle extends beyond the
throttle_threshold, for the num_scaling periods of time, over provision
the quota. The the quota is scaled up by the scale_up_factor when
throttle is observed and scaled down by the scale_down_factor when not.
No other tracing and recommendation is performed during this time.

Signed-off-by: Pratik Rajesh Sampat <psampat2@illinois.edu>