trace: kscaler: reduce history from 100 to 10
When we define structures for ebpf trace, size of 100 is too large and
causes issues. So defined it to 10 for now. Maybe that's enough, not
sure. We will have to trail and error this out or find out a size
agnostic way to deal with arrays which is hard in ebpf.

Signed-off-by: Pratik Rajesh Sampat <psampat2@illinois.edu>